feat(Grup B): B1-B6 - labels, cantitate decimala, excel, anulare, validare stoc
- B1: Data livrare → Termen livrare in cosul comenzii noi
- B2: Coloana Produse → Cantitate in Comenzile mele si Toate Comenzile
- B3: Filename Excel → client-data-#nr.csv (unic si identificabil)
- B4: parseInt → parseFloat + step=0.01 pe toate inputurile de cantitate
- B5: Buton Anuleaza adaugat in Comenzile mele (la fel ca in Toate Comenzile)
- B6: Input cantitate rosu cand depaseste stocul disponibil

+++ src/app/features/transport/transport.component.ts
@@ -415,7 +415,7 @@ export class TransportComponent implements OnInit {
     const order = this.modalOrders().find(o => o.id === orderId);
     if (!order) return;
     const max = this.getRemainingQtyArr(order, this.editingId() ?? undefined)[idx] ?? 0;
-    const qty = Math.min(max, Math.max(0, parseInt(String(val)) || 0));
+    const qty = Math.min(max, Math.max(0, parseFloat(String(val)) || 0));
     this.modalQty.update(m => ({ ...m, [orderId]: { ...(m[orderId] ?? {}), [idx]: qty } }));
   }
